backend/requirements.txt
Flask
neo4j
langchain
langchain-community
flask-cors
python-dotenv
langchain-groq